wip - forest textures

+++ include/util.h
@@ -265,7 +265,7 @@ inline bool readFile(const std::string &path, T& content, bool quiet = false)
 }
 
 
-inline bool writeFile(const std::string &path, const char *data, size_t data_size)
+inline bool writeFile(const std::string &path, const char *data, int data_size)
 {
   using namespace std;
